Updated Maven deps in effort to enable pdf export.
Building the jetty image using Make:

```bash
make docker-build
```

yields a Docker image that support eps/png/svg export but not pdf.
I added all of the dependencies mentioned in the
[documentation](https://plantuml.com/pdf) as requirements for pdf
support. However, when I attempt to export a pdf, I receive a
'404 Not Found' error response from the server.